Support Mill 0.11.0-M11
- Drop support for Mill 0.10
