--- itest/0.11/src/simple/core/src/main.scala
@@ -1,3 +1,0 @@
-object core {
-  val core = 1
-}
