--- itest/0.11/src/simple/no-native/src/main.scala
@@ -1,5 +1,0 @@
-object main {
-  def main(args: Array[String]): Unit = {
-    print(other.other + 1)
-  }
-}
