--- itest/0.11/src/simple/other/src/main.scala
@@ -1,3 +1,0 @@
-object other {
-  val other = core.core + 1
-}
