+++ itest/src/simple/build.sc
@@ -2,7 +2,6 @@ import $file.plugins
 
 import mill._
 import mill.api.Result
-import mill.define.SelectMode
 import mill.eval._
 import mill.main.MainModule
 import mill.scalalib._
@@ -58,84 +57,63 @@ trait CrossScalaJSNative extends CrossPlatform {
   object jvm extends Shared with CommonJVM
   object js extends Cross[JS]("1.13.1")
   trait JS extends Shared with CrossScalaJSModule
-  object native extends Cross[Native]("0.4.12")
+  object native extends Cross[Native]("0.4.14")
   trait Native extends Shared with CrossScalaNativeModule
-}
-
-def execute(ev: Evaluator, command: String)(f: ujson.Value => Unit) = {
-  MainModule.evaluateTasks(
-    ev,
-    Seq(command),
-    SelectMode.Single
-  )(v => f(v.head._2.arr.head))
 }
 
 def verify(ev: Evaluator) = T.command {
   `no-native`("3.2.2").jvm.compile()
 
   locally {
-    val Result.Success(_) = execute(ev, "core[3.2.2].jvm.artifactName") {
-      case ujson.Str(result) =>
-        assert(
-          result == "core",
-          s"Wrong artifactName: $result"
-        )
-    }
+    val result = ev.evalOrThrow()(core("3.2.2").jvm.artifactName)
+    assert(
+      result == "core",
+      s"Wrong artifactName: $result"
+    )
   }
 
   locally {
-    val Result.Success(_) =
-      execute(ev, "crossScalaJSNative[3.2.2].native[0.4.12].artifactName") {
-        case ujson.Str(result) =>
-          assert(
-            result == "crossScalaJSNative",
-            s"Wrong artifactName: $result"
-          )
-      }
+    val result = ev.evalOrThrow()(
+      crossScalaJSNative("3.2.2").native("0.4.14").artifactName
+    )
+    assert(
+      result == "crossScalaJSNative",
+      s"Wrong artifactName: $result"
+    )
   }
 
   locally {
-    val Result.Success(_) =
-      execute(ev, "crossScalaJSNative[3.2.2].js[1.13.1].artifactName") {
-        case ujson.Str(result) =>
-          assert(
-            result == "crossScalaJSNative",
-            s"Wrong artifactName: $result"
-          )
-      }
+    val result =
+      ev.evalOrThrow()(crossScalaJSNative("3.2.2").js("1.13.1").artifactName)
+    assert(
+      result == "crossScalaJSNative",
+      s"Wrong artifactName: $result"
+    )
   }
 
   locally {
-    val Result.Success(_) =
-      execute(ev, "crossScalaJSNative[3.2.2].js[1.13.1].scalaVersion") {
-        case ujson.Str(result) =>
-          assert(
-            result == "3.2.2",
-            s"Wrong scalaVersion: $result"
-          )
-      }
+    val result =
+      ev.evalOrThrow()(crossScalaJSNative("3.2.2").js("1.13.1").scalaVersion)
+    assert(
+      result == "3.2.2",
+      s"Wrong artifactName: $result"
+    )
   }
 
   locally {
-    val Result.Success(_) =
-      execute(ev, "normal.jvm.sources") { case ujson.Arr(arr) =>
-        val path = os.Path(arr.head.str.split(':').last) / os.up
-        assert(
-          path.last == "normal",
-          s"Wrong millSourcePath: $path"
-        )
-      }
+    val path = ev.evalOrThrow()(normal.jvm.sources).head.path / os.up
+    assert(
+      path.last == "normal",
+      s"Wrong millSourcePath: $path"
+    )
   }
 
   locally {
-    val Result.Success(_) =
-      execute(ev, "core[3.2.2].jvm.sources") { case ujson.Arr(arr) =>
-        val path = os.Path(arr.head.str.split(':').last) / os.up
-        assert(
-          path.last == "core",
-          s"Wrong millSourcePath: $path"
-        )
-      }
+    val path = ev.evalOrThrow()(core("3.2.2").jvm.sources).head.path / os.up
+    assert(
+      path.last == "core",
+      s"Wrong millSourcePath: $path"
+    )
   }
 
   ()
