+++ mill-crossplatform/src/com/github/lolgab/mill/crossplatform/CrossPlatform.scala
@@ -1,6 +1,7 @@
 package com.github.lolgab.mill.crossplatform
 
 import mill._
+import mill.main.BuildInfo
 import mill.define.Cross
 import mill.scalajslib._
 import mill.scalalib._
@@ -8,29 +9,37 @@ import mill.scalanativelib._
 
 import scala.language.reflectiveCalls
 
-trait CrossPlatform extends Module with VersionSpecific.CrossPlatform {
+trait CrossPlatform extends Cross.Module[String] {
   container =>
+  override def crossValue: String = throw new Exception(
+    """CrossPlatform defines `def crossValue: String` only because it's required by Cross.Module[String].
+      |You shouldn't need to use it directly. If you see this message please report a bug at https://github.com/lolgab/mill-crossplatform/issues/new""".stripMargin
+  )
+
+  private[crossplatform] protected def myArtifactNameParts: Seq[String] =
+    millModuleSegments.parts.init
+
   CrossPlatform.checkMillVersion()
   def moduleDeps: Seq[CrossPlatform] = Seq.empty
   def compileModuleDeps: Seq[CrossPlatform] = Seq.empty
 
-  def enableJVM: Boolean = true
-  def enableJS: Boolean = true
-  def enableNative: Boolean = true
+  private def enableJVM: Boolean = true
+  private def enableJS: Boolean = true
+  private def enableNative: Boolean = true
 
   private def platforms: Seq[String] = {
     def loop(module: Module): Seq[String] = module match {
       case _: ScalaNativeModule => Seq("native")
       case _: ScalaJSModule     => Seq("js")
       case _: ScalaModule       => Seq("jvm")
-      case VersionSpecific.IsCross(c) =>
+      case c: Cross[_] =>
         loop(
-          VersionSpecific.getModules(c).head.asInstanceOf[Module]
+          c.crossModules.head.asInstanceOf[Module]
         )
       case _ => Seq.empty[String]
     }
     millInternal
-      .reflectNestedObjects[Module]
+      .reflectNestedObjects[Module]()
       .flatMap(m => loop(m))
       .toSeq
   }
@@ -39,24 +48,19 @@ trait CrossPlatform extends Module with VersionSpecific.CrossPlatform {
     case _: ScalaNativeModule => enableNative
     case _: ScalaJSModule     => enableJS
     case _: ScalaModule       => enableJVM
-    case VersionSpecific.IsCross(c) =>
+    case c: Cross[_] =>
       enableModuleCondition(
-        VersionSpecific.getModules(c).head.asInstanceOf[Module]
+        c.crossModules.head.asInstanceOf[Module]
       )
     case _ => true
   }
-  override lazy val millModuleDirectChildren: Seq[Module] =
-    millInternal
-      .reflectNestedObjects[Module]
+  override def millModuleDirectChildren: Seq[Module] =
+    super.millModuleDirectChildren
       .filter(enableModuleCondition)
-      .toSeq
 
-  trait CrossPlatformScalaModule
-      extends ScalaModule
-      with VersionSpecific.CrossPlatformScalaModule {
+  trait CrossPlatformScalaModule extends ScalaModule {
+    override def artifactNameParts = container.myArtifactNameParts
     override def millSourcePath = super.millSourcePath / os.up
-    private[crossplatform] protected def myArtifactNameParts: Seq[String] =
-      container.myArtifactNameParts
     override def moduleDeps = super.moduleDeps ++
       container.moduleDeps.map(innerModule).asInstanceOf[Seq[this.type]]
     override def compileModuleDeps = super.compileModuleDeps ++
@@ -82,7 +86,7 @@ trait CrossPlatform extends Module with VersionSpecific.CrossPlatform {
       super.sources() ++ platformSources(millSourcePath)
     }
 
-    trait CrossPlatformSources extends Tests {
+    trait CrossPlatformSources extends ScalaModuleTests {
       override def sources = T.sources {
         super.sources() ++ platformSources(millSourcePath)
       }
@@ -131,23 +135,21 @@ trait CrossPlatform extends Module with VersionSpecific.CrossPlatform {
   implicit def crossPlatformResolver: Cross.Resolver[CrossPlatform] =
     new Cross.Resolver[CrossPlatform] {
       def resolve[V <: CrossPlatform](c: Cross[V]): V = {
-        val scalaV = myCrossValue
+        val scalaV = crossValue
         scalaV
           .split('.')
           .inits
           .takeWhile(_.length > 1)
           .flatMap(prefix =>
-            VersionSpecific
-              .getModules(c)
-              .find(_.myCrossValue.split('.').startsWith(prefix))
+            c.crossModules
+              .find(_.crossValue.split('.').startsWith(prefix))
           )
           .collectFirst { case x => x }
           .getOrElse(
             throw new Exception(
               s"Unable to find compatible cross version between $scalaV and " +
-                VersionSpecific
-                  .getModules(c)
-                  .map(_.myCrossValue)
+                c.crossModules
+                  .map(_.crossValue)
                   .mkString(",")
             )
           )
@@ -156,10 +158,14 @@ trait CrossPlatform extends Module with VersionSpecific.CrossPlatform {
 
   trait CrossPlatformCrossScalaModule
       extends CrossScalaModule
-      with CrossPlatformScalaModule
-      with VersionSpecific.CrossPlatformCrossScalaModule {
+      with CrossPlatformScalaModule {
+    override def crossValue: String = throw new Exception(
+      """CrossPlatformCrossScalaModule defines `def crossValue: String` only because it's required by CrossScalaModule (defined in Cross.Module[String]).
+      |You shouldn't need to use it directly. If you see this message please report a bug at https://github.com/lolgab/mill-crossplatform/issues/new""".stripMargin
+    )
+    override def crossScalaVersion: String = myCrossValue
     private[crossplatform] protected def myCrossValue: String =
-      container.myCrossValue
+      container.crossValue
   }
 }
 object CrossPlatform {
