+++ mill-crossplatform/src/com/github/lolgab/mill/crossplatform/CrossScalaJSModule.scala
@@ -3,8 +3,7 @@ package com.github.lolgab.mill.crossplatform
 import mill._
 import mill.scalajslib._
 
-trait CrossScalaJSModule
-    extends ScalaJSModule
-    with VersionSpecific.CrossScalaJSModule {
+trait CrossScalaJSModule extends ScalaJSModule with Cross.Module[String] {
+  def crossScalaJSVersion: String = crossValue
   override def scalaJSVersion = crossScalaJSVersion
 }
